Refactored the Odometer out into it's own file, Updated Read Me
